ref(replays): move start time in replay details
